Fix incorrect assumption on output directory name.
While executing the extract command, dpkg-source by default extracts the
source package to an output directory named as source-version. If the
source package format is 3.0 (quilt), the version string is taken from
the orig tarball, not the dsc file. Baiji attempts to use the
dsc version string, and fails if no correspondingly named output
directory is found after the extraction.