Don't check build dependencies in the host system.
By default dpkg-buildpackage checks build dependencies and conflicts
before building the package, no matter whether source or binary. When no
.dsc file is supplied to the build command, baiji uses dpkg-buildpackage
to build a source package from the current directory first. This is
liable to fail if the host system doesn't satisfy the build
dependencies. Since the actual building job is done inside a Docker
environment, we don't need this behavior.